pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.1"
description = "It's tool to compare two configuration files and show differences."
authors = [
    { name = "vlrkors", email = "vlrkors@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0"
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.11.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[project.scripts]
gendiff = "gendiff.scripts.cli:main"

[tool.ruff]
line-length = 80
select = ["E", "F", "W", "I"]
ignore = []

[tool.ruff.format]
docstring-code-format = true

[tool.black]
line-length = 80
target-version = ["py312"]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black",
    "pytest-cov"
]

